refactor: Move pkg/mcpserver to pkg/framework/mcpserver
Relocate MCP server implementation to framework package to:
- Avoid naming collision with mark3labs/mcp-go/mcp
- Establish as part of public framework API
- Align with installer framework architecture plan

File history preserved via git mv.
